pyproject.toml
[project]
name = "dictator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "assemblyai>=0.40.2",
    "deepgram-sdk>=4.1.0",
    "litellm>=1.70.4",
    "pillow>=11.2.1",
    "pyaudio>=0.2.14",
    "pystray>=0.19.5",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "xdotool>=0.4.0",
]
